Extract DeepSeekV2 attention gate helpers
